update rgb-lib to 0.3.0-alpha.13

+++ app/src/main/java/com/iriswallet/ui/MainViewModel.kt
@@ -262,7 +262,7 @@ class MainViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel(
         recipient: String,
         amount: String,
         transportEndpoints: List<String>,
-        feeRate: Float,
+        feeRate: ULong,
     ) {
         tryCallWithTimeout(AppConstants.longTimeout, _sent) {
             val txid =
